requirements.txt
solidAPI==0.1.6
pafy==0.5.5
python-dotenv==0.19.1
youtube-dl==2021.6.6
youtube-search==2.1.0
git+https://github.com/pyrogram/pyrogram@master
py-tgcalls==0.8.1rc1
aiohttp==3.7.4.post0